Point at toml-test
We could push toml-test over this (maybe even force-push so history is
preserved), but then we'd have two repos and that's confusing.

The best case would be to move/rename BurntSushi/toml-test to here; I
*think* GitHub will then redirect BurntSushi/toml-test, but not 100%
sure about that, and if it will remain working for "deep links" to files
or issue comments etc.

Also, there's the issue with repo/organisation permissions.

Anyway, to avoid confusion and double work (like #8) just point to the
toml-test repo in the README. I think that's probably also the best
long-term solution; that repo has been around for a decade, and renaming
things tends to cause more problems than it solves.

Closes #8